+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the library will be documented in this file.
 ## vX.X.X (Month DD, YYYY)
 
 - Updated the README.md file.
+- Added `borderRadius` shorthand.
 
 ## v0.2.1 (August 19, 2023)